Add logger and update build function

+++ functions/build.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 npm run build
 
+# check .build folder exists
+if [ ! -d ".build" ]; then
 rm -rf .build
-rm -rf temp
+fi
 mkdir .build
+# check temp folder exists
+if [ ! -d "temp" ]; then
+rm -rf temp
+fi
 mkdir temp
 cd .build
 mkdir src
@@ -11,4 +17,9 @@ cp -r ../dist/* ./src/
 cp ../package.json ./
 
 npm install --omit=dev
-c:/Program\ Files/7-Zip/7z.exe a ../temp/lambda.zip *
+# check if zip exists and using zip to zip up lambda else using 7z
+if [ -f "c:/Program Files/7-Zip/7z.exe" ]; then
+    c:/Program\ Files/7-Zip/7z.exe a ../temp/lambda.zip *
+else 
+    zip -r ../temp/lambda.zip *
+fi
